+++ endpoints/skills/routers.py
@@ -39,6 +39,10 @@ def create_a_new_skill(skill: Skill, request: Request):
 # GET a skill by skill type
 @router.get('/{technology}', response_model=Skill)
 def get_a_skill_by_technology(technology: str, request: Request):
+    # check if technology exists in MongoDB db
+    if request.app.db['skills'].find({"technology": technology}, {"$exists": False}):
+        raise HTTPException(400, f"{technology} skill does not exist!")
+
     oneSkill = request.app.db['skills'].find_one({"technology": technology})
 
     response = json.loads(json.dumps(oneSkill, default=json_util.default))
@@ -50,6 +54,10 @@ def get_a_skill_by_technology(technology: str, request: Request):
 # UPDATE a skill by skill type
 @router.put('/{technology}', response_model=UpdateSkill)
 def update_a_skill_by_technology(technology: str, skill: Skill, request: Request):
+    # check if technology exists in MongoDB db
+    if request.app.db['skills'].find({"technology": technology}, {"$exists": False}):
+        raise HTTPException(400, f"{technology} skill does not exist!")
+
     updatedSkill = skill.dict()
     request.app.db['skills'].update_one({"technology": technology}, { "$set": updatedSkill})
     # converts list object to str (JSON format)
@@ -63,6 +71,10 @@ def update_a_skill_by_technology(technology: str, skill: Skill, request: Request
 # DELETE a skill by skill type
 @router.delete('/{technology}', response_model=Skill)
 def delete_a_skill_by_technology(technology: str, request: Request):
+    # check if technology exists in MongoDB db
+    if request.app.db['skills'].find({"technology": technology}, {"$exists": False}):
+        raise HTTPException(400, f"{technology} skill does not exist!")
+
     deletedSkill = request.app.db["skills"].delete_one({"technology": technology})
 
     if deletedSkill.deleted_count == 1:
